docker-chrome-headless#4: Remove xfonts-cyrillic, ttf-wqy-zenhei and fonts-unfonts-core from build
This removes Chinese, Cyrillic and Korean fonts from the build, with my
apologies to anyone affected; an upstream change meant that the
inclusion of Cyrillic fonts broke builds as of Debian 12, and I'm
proactively reducing maintenance surface area by removing the others.

Please use or build a Chrome image which provides support for the fonts
you require.